runtime: massive breaking restructuring
All runtime models are now in their own namespace,

+++ components/runtime-models/gen-move-types.bash
@@ -11,5 +11,7 @@ cargo t
 ./gen-index.bash
 
 # move bindings
-rm -r ../runtime/src/ts/models
-mv bindings/ ../runtime/src/ts/models
+if [ -d "../runtime/src/ts/models" ]; then 
+    rm -r ../runtime/src/ts/models
+fi
+mv bindings/ ../runtime/src/ts/models
